# src/docs/management/management_swagger.yaml
basePath: /api/v1
definitions:
  api.ActivateRequest:
    properties:
      agents:
        additionalProperties:
          type: string
        type: object
    type: object
...
  api.WorkspaceRequest:
    properties:
      cacheName:
        type: string
      name:
        type: string
    required:
    - cacheName
    - name
    type: object
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  model.Agent:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      name:
        type: string
      token:
        type: string
      updatedAt:
        type: string
      uuid:
        type: string
      workspace_id:
        type: integer
    type: object
  model.BinaryCache:
    properties:
      access:
        $ref: '#/definitions/model.BinaryCacheAccess'
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      name:
        type: string
      port:
        type: integer
      publicKey:
        type: string
      retention:
        type: integer
      store_paths:
        items:
          $ref: '#/definitions/model.StorePath'
        type: array
      token:
        type: string
      updatedAt:
        type: string
      url:
        type: string
      uuid:
        type: string
      workspaces:
        items:
          $ref: '#/definitions/model.Workspace'
        type: array
    type: object
  model.BinaryCacheAccess:
    enum:
    - private
    - public
    type: string
    x-enum-varnames:
    - Private
    - Public
  model.Deployment:
    properties:
      agent:
        $ref: '#/definitions/model.Agent'
      agentId:
        type: integer
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: string
      status:
        $ref: '#/definitions/model.DeploymentStatus'
      storePath:
        type: string
      updatedAt:
        type: string
    type: object
  model.DeploymentStatus:
    enum:
    - pending
    - in_progress
    - success
    - failed
    type: string
    x-enum-varnames:
    - DeploymentPending
    - DeploymentInProgress
    - DeploymentSuccess
    - DeploymentFailed
  model.ErrorResponse:
    properties:
      additional_info:
        items:
          type: string
        type: array
      error:
        example: not found
        type: string
    type: object
  model.StorePath:
    properties:
      binaryCacheId:
        description: |-
          BinaryCacheId uint         `json:"binarycacheId"`
          BinaryCache   *BinaryCache `json:"-"`
        type: integer
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      deriver:
        type: string
      fileHash:
        type: string
      fileSize:
        type: integer
      id:
        type: integer
      narHash:
        type: string
      narSize:
        type: integer
      references:
        description: Stored as a string (serialized)
        type: string
      storeHash:
        type: string
      storeSuffix:
        type: string
      updatedAt:
        type: string
    type: object
  model.Workspace:
    properties:
      agents:
        items:
          $ref: '#/definitions/model.Agent'
        type: array
      binary_cache:
        $ref: '#/definitions/model.BinaryCache'
      binary_cache_id:
        type: integer
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      name:
        type: string
      token:
        type: string
      updatedAt:
        type: string
    type: object
info:
  contact: {}
  description: API for managing binary caches, workspaces, and agents.
  title: Management API
  version: "1.0"
paths:
...
  /deploy/activate:
    post:
      consumes:
      - application/json
      description: Activate deployments for multiple agents.
      parameters:
      - description: Activation details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.ActivateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Deployment'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Activate deployment
      tags:
      - deployment
  /deploy/agent/{workspace}/{name}:
    delete:
      description: Delete an agent from a workspace.
      parameters:
      - description: Workspace Name
        in: path
        name: workspace
        required: true
        type: string
      - description: Agent Name
        in: path
        name: name
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Delete agent
      tags:
      - agent
    get:
      description: Get detailed information about an agent in a workspace.
      parameters:
      - description: Workspace Name
        in: path
        name: workspace
        required: true
        type: string
      - description: Agent Name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Agent'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Get agent info
      tags:
      - agent
    post:
      description: Create a new agent in a workspace.
      parameters:
      - description: Workspace Name
        in: path
        name: workspace
        required: true
        type: string
      - description: Agent Name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Agent'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Create agent
      tags:
      - agent
  /deploy/deployment/{workspace}:
    get:
      description: Get detailed information about a deployment by UUID.
      parameters:
      - description: Deployment UUID
        in: path
        name: workspace
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Deployment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Get deployment info
      tags:
      - deployment
  /deploy/deployment/{workspace}/{name}:
    get:
      description: List all deployments for a specific agent in a workspace.
      parameters:
      - description: Workspace Name
        in: path
        name: workspace
        required: true
        type: string
      - description: Agent Name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Deployment'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Get deployments for agent
      tags:
      - deployment
    post:
      description: Create a new deployment for an agent.
      parameters:
      - description: Workspace Name
        in: path
        name: workspace
        required: true
        type: string
      - description: Agent Name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Deployment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Create deployment
      tags:
      - deployment
  /deploy/deployment/{workspace}/{name}/{index}:
    get:
      description: Get a specific deployment for an agent by its index.
      parameters:
      - description: Workspace Name
        in: path
        name: workspace
        required: true
        type: string
      - description: Agent Name
        in: path
        name: name
        required: true
        type: string
      - description: Deployment Index
        in: path
        name: index
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Get deployment by index
      tags:
      - deployment
...
  /deploy/workspace:
    post:
      consumes:
      - application/json
      description: Create a new workspace associated with a binary cache.
      parameters:
      - description: Workspace details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.WorkspaceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Workspace'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Create workspace
      tags:
      - workspace
  /deploy/workspace/{workspace}:
    delete:
      description: Delete a workspace by name.
      parameters:
      - description: Workspace Name
        in: path
        name: workspace
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Delete workspace
      tags:
      - workspace
    get:
      description: Get detailed information about a workspace.
      parameters:
      - description: Workspace Name
        in: path
        name: workspace
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Workspace'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Get workspace info
      tags:
      - workspace
  /deploy/workspace/{workspace}/agents:
    get:
      description: List all agents in a workspace.
      parameters:
      - description: Workspace Name
        in: path
        name: workspace
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Agent'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: List agents
      tags:
      - workspace
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"