# src/model/agent.go
package model

import "github.com/google/uuid"

// Agent represents a deployment agent belonging to a workspace.
type Agent struct {
	ID    uint      `gorm:"primarykey"`
	Uuid  uuid.UUID `gorm:"type:uuid;unique;not null"`
	Name  string    `gorm:"type:varchar(100);unique;not null" json:"name"`
	Token string    `gorm:"type:text;not null" json:"-"` // not serialized agent token

	WorkspaceId uint       `json:"workspace_id"`
	Workspace   *Workspace `gorm:"foreignKey:WorkspaceId;constraint:OnUpdate:CASCADE;" json:"-"`
}